libs/project: Call functions from any project
Helper functions, such as "arguments", provided by one project should
be callable from another project for more flexibility. Metaprojects in
particular could use this functionality.

In order for this to work, project_include() is used to create the
necessary environment separation (so defined functions are not
clobbered by the helper function).