Create new meta target working with dependencies
The specified action will be taken on each of the dependencies
returned by project_dependencies() where $project is the first
argument to the meta-target, e.g.

./libreboot download libreboot-dependencies coreboot x200

The above command would download only the minimum projects necessary
for building all x200 image configurations.

To then build those dependencies:

./libreboot build libreboot-dependencies coreboot x200

And to get the final images:

./libreboot build coreboot x200