Remove dependencies_check() from project scripts
Removing the check function allows dependencies() to be executed,
printing the list of dependencies. It was not executed before due to
the way check functions are handled; check functions are executed
before functions with the same name (minus the "_check" suffix) in
order to determine whether calling the function is necessary.